refactor: move DI composition root from core/ to app/
The generated injectable config (`injection.config.dart`) imports every
feature's data/domain/presentation, so housing it under `core/` made
`core` depend on `features` — inverting the intended
`features -> shared -> ui -> core` dependency direction.

Move `lib/core/di/` to `lib/app/di/`. The app shell is the composition
root and is explicitly allowed to depend on features, so the wiring now
sits where it belongs and `core/` no longer imports any feature. Updates
all `getIt`/`configureDependencies` import sites (11 lib files, 2 tests)
and regenerates the config in its new location.
