+++ lib/features/auth/presentation/screens/change_password_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-import '../../../../core/di/injection.dart';
+import '../../../../app/di/injection.dart';
 import '../../../../core/extensions/build_context_extensions.dart';
 import '../bloc/change_password_cubit.dart';
 import '../bloc/change_password_state.dart';
