+++ lib/features/bookmarks/presentation/screens/bookmarks_list_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../core/di/injection.dart';
+import '../../../../app/di/injection.dart';
 import '../bloc/bookmarks_list/bookmarks_list_bloc.dart';
 import '../widgets/bookmarks_list_widgets.dart';
